Create decimal time from time vector

Create decimal time on an annual scale from an input time vector

dec_time(date_in)

# S3 method for class 'Date'
dec_time(date_in)

Arguments

date_in

input time vector, usually a date object

Value

A named list of four numeric vectors including day_num (decimal day on an annual scale), month (month of the year as integer), year, and dec_time (decimal time as sum of year and day_num)

Details

Function is used internally within the package.

Examples

dt <- Sys.Date()
dts <- seq.Date(dt - 365, dt, by = 'day') 

dec_time(dts)
#> $day_num
#>   [1] 0.043835616 0.046575342 0.049315068 0.052054795 0.054794521 0.057534247
#>   [7] 0.060273973 0.063013699 0.065753425 0.068493151 0.071232877 0.073972603
#>  [13] 0.076712329 0.079452055 0.082191781 0.084931507 0.087671233 0.090410959
#>  [19] 0.093150685 0.095890411 0.098630137 0.101369863 0.104109589 0.106849315
#>  [25] 0.109589041 0.112328767 0.115068493 0.117808219 0.120547945 0.123287671
#>  [31] 0.126027397 0.128767123 0.131506849 0.134246575 0.136986301 0.139726027
#>  [37] 0.142465753 0.145205479 0.147945205 0.150684932 0.153424658 0.156164384
#>  [43] 0.158904110 0.161643836 0.164383562 0.164383562 0.169863014 0.172602740
#>  [49] 0.175342466 0.178082192 0.180821918 0.183561644 0.186301370 0.189041096
#>  [55] 0.191780822 0.194520548 0.197260274 0.200000000 0.202739726 0.205479452
#>  [61] 0.208219178 0.210958904 0.213698630 0.216438356 0.219178082 0.221917808
#>  [67] 0.224657534 0.227397260 0.230136986 0.232876712 0.235616438 0.238356164
#>  [73] 0.241095890 0.243835616 0.246575342 0.249315068 0.249315068 0.254794521
#>  [79] 0.257534247 0.260273973 0.263013699 0.265753425 0.268493151 0.271232877
#>  [85] 0.273972603 0.276712329 0.279452055 0.282191781 0.284931507 0.287671233
#>  [91] 0.290410959 0.293150685 0.295890411 0.298630137 0.301369863 0.304109589
#>  [97] 0.306849315 0.309589041 0.312328767 0.315068493 0.317808219 0.320547945
#> [103] 0.323287671 0.326027397 0.328767123 0.331506849 0.331506849 0.336986301
#> [109] 0.339726027 0.342465753 0.345205479 0.347945205 0.350684932 0.353424658
#> [115] 0.356164384 0.358904110 0.361643836 0.364383562 0.367123288 0.369863014
#> [121] 0.372602740 0.375342466 0.378082192 0.380821918 0.383561644 0.386301370
#> [127] 0.389041096 0.391780822 0.394520548 0.397260274 0.400000000 0.402739726
#> [133] 0.405479452 0.408219178 0.410958904 0.413698630 0.416438356 0.416438356
#> [139] 0.421917808 0.424657534 0.427397260 0.430136986 0.432876712 0.435616438
#> [145] 0.438356164 0.441095890 0.443835616 0.446575342 0.449315068 0.452054795
#> [151] 0.454794521 0.457534247 0.460273973 0.463013699 0.465753425 0.468493151
#> [157] 0.471232877 0.473972603 0.476712329 0.479452055 0.482191781 0.484931507
#> [163] 0.487671233 0.490410959 0.493150685 0.495890411 0.498630137 0.498630137
#> [169] 0.504109589 0.506849315 0.509589041 0.512328767 0.515068493 0.517808219
#> [175] 0.520547945 0.523287671 0.526027397 0.528767123 0.531506849 0.534246575
#> [181] 0.536986301 0.539726027 0.542465753 0.545205479 0.547945205 0.550684932
#> [187] 0.553424658 0.556164384 0.558904110 0.561643836 0.564383562 0.567123288
#> [193] 0.569863014 0.572602740 0.575342466 0.578082192 0.580821918 0.583561644
#> [199] 0.583561644 0.589041096 0.591780822 0.594520548 0.597260274 0.600000000
#> [205] 0.602739726 0.605479452 0.608219178 0.610958904 0.613698630 0.616438356
#> [211] 0.619178082 0.621917808 0.624657534 0.627397260 0.630136986 0.632876712
#> [217] 0.635616438 0.638356164 0.641095890 0.643835616 0.646575342 0.649315068
#> [223] 0.652054795 0.654794521 0.657534247 0.660273973 0.663013699 0.665753425
#> [229] 0.668493151 0.668493151 0.673972603 0.676712329 0.679452055 0.682191781
#> [235] 0.684931507 0.687671233 0.690410959 0.693150685 0.695890411 0.698630137
#> [241] 0.701369863 0.704109589 0.706849315 0.709589041 0.712328767 0.715068493
#> [247] 0.717808219 0.720547945 0.723287671 0.726027397 0.728767123 0.731506849
#> [253] 0.734246575 0.736986301 0.739726027 0.742465753 0.745205479 0.747945205
#> [259] 0.750684932 0.750684932 0.756164384 0.758904110 0.761643836 0.764383562
#> [265] 0.767123288 0.769863014 0.772602740 0.775342466 0.778082192 0.780821918
#> [271] 0.783561644 0.786301370 0.789041096 0.791780822 0.794520548 0.797260274
#> [277] 0.800000000 0.802739726 0.805479452 0.808219178 0.810958904 0.813698630
#> [283] 0.816438356 0.819178082 0.821917808 0.824657534 0.827397260 0.830136986
#> [289] 0.832876712 0.835616438 0.835616438 0.841095890 0.843835616 0.846575342
#> [295] 0.849315068 0.852054795 0.854794521 0.857534247 0.860273973 0.863013699
#> [301] 0.865753425 0.868493151 0.871232877 0.873972603 0.876712329 0.879452055
#> [307] 0.882191781 0.884931507 0.887671233 0.890410959 0.893150685 0.895890411
#> [313] 0.898630137 0.901369863 0.904109589 0.906849315 0.909589041 0.912328767
#> [319] 0.915068493 0.917808219 0.917808219 0.923287671 0.926027397 0.928767123
#> [325] 0.931506849 0.934246575 0.936986301 0.939726027 0.942465753 0.945205479
#> [331] 0.947945205 0.950684932 0.953424658 0.956164384 0.958904110 0.961643836
#> [337] 0.964383562 0.967123288 0.969863014 0.972602740 0.975342466 0.978082192
#> [343] 0.980821918 0.983561644 0.986301370 0.989041096 0.991780822 0.994520548
#> [349] 0.997260274 1.000000000 1.002739726 0.005479452 0.008219178 0.010958904
#> [355] 0.013698630 0.016438356 0.019178082 0.021917808 0.024657534 0.027397260
#> [361] 0.030136986 0.032876712 0.035616438 0.038356164 0.041095890 0.043835616
#> 
#> $month
#>   [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2
#>  [26]  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3
#>  [51]  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3
#>  [76]  3  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4
#> [101]  4  4  4  4  4  4  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5
#> [126]  5  5  5  5  5  5  5  5  5  5  5  5  6  6  6  6  6  6  6  6  6  6  6  6  6
#> [151]  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7  7
#> [176]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  8  8
#> [201]  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8
#> [226]  8  8  8  8  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9
#> [251]  9  9  9  9  9  9  9  9  9 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#> [276] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 11 11 11 11 11 11 11 11 11 11
#> [301] 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 12 12 12 12 12
#> [326] 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12
#> [351] 12  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> $year
#>   [1] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#>  [16] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#>  [31] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#>  [46] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#>  [61] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#>  [76] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#>  [91] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [106] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [121] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [136] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [151] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [166] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [181] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [196] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [211] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [226] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [241] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [256] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [271] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [286] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [301] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [316] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [331] 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025 2025
#> [346] 2025 2025 2025 2025 2025 2025 2026 2026 2026 2026 2026 2026 2026 2026 2026
#> [361] 2026 2026 2026 2026 2026 2026
#> 
#> $dec_time
#>   [1] 2025.044 2025.047 2025.049 2025.052 2025.055 2025.058 2025.060 2025.063
#>   [9] 2025.066 2025.068 2025.071 2025.074 2025.077 2025.079 2025.082 2025.085
#>  [17] 2025.088 2025.090 2025.093 2025.096 2025.099 2025.101 2025.104 2025.107
#>  [25] 2025.110 2025.112 2025.115 2025.118 2025.121 2025.123 2025.126 2025.129
#>  [33] 2025.132 2025.134 2025.137 2025.140 2025.142 2025.145 2025.148 2025.151
#>  [41] 2025.153 2025.156 2025.159 2025.162 2025.164 2025.164 2025.170 2025.173
#>  [49] 2025.175 2025.178 2025.181 2025.184 2025.186 2025.189 2025.192 2025.195
#>  [57] 2025.197 2025.200 2025.203 2025.205 2025.208 2025.211 2025.214 2025.216
#>  [65] 2025.219 2025.222 2025.225 2025.227 2025.230 2025.233 2025.236 2025.238
#>  [73] 2025.241 2025.244 2025.247 2025.249 2025.249 2025.255 2025.258 2025.260
#>  [81] 2025.263 2025.266 2025.268 2025.271 2025.274 2025.277 2025.279 2025.282
#>  [89] 2025.285 2025.288 2025.290 2025.293 2025.296 2025.299 2025.301 2025.304
#>  [97] 2025.307 2025.310 2025.312 2025.315 2025.318 2025.321 2025.323 2025.326
#> [105] 2025.329 2025.332 2025.332 2025.337 2025.340 2025.342 2025.345 2025.348
#> [113] 2025.351 2025.353 2025.356 2025.359 2025.362 2025.364 2025.367 2025.370
#> [121] 2025.373 2025.375 2025.378 2025.381 2025.384 2025.386 2025.389 2025.392
#> [129] 2025.395 2025.397 2025.400 2025.403 2025.405 2025.408 2025.411 2025.414
#> [137] 2025.416 2025.416 2025.422 2025.425 2025.427 2025.430 2025.433 2025.436
#> [145] 2025.438 2025.441 2025.444 2025.447 2025.449 2025.452 2025.455 2025.458
#> [153] 2025.460 2025.463 2025.466 2025.468 2025.471 2025.474 2025.477 2025.479
#> [161] 2025.482 2025.485 2025.488 2025.490 2025.493 2025.496 2025.499 2025.499
#> [169] 2025.504 2025.507 2025.510 2025.512 2025.515 2025.518 2025.521 2025.523
#> [177] 2025.526 2025.529 2025.532 2025.534 2025.537 2025.540 2025.542 2025.545
#> [185] 2025.548 2025.551 2025.553 2025.556 2025.559 2025.562 2025.564 2025.567
#> [193] 2025.570 2025.573 2025.575 2025.578 2025.581 2025.584 2025.584 2025.589
#> [201] 2025.592 2025.595 2025.597 2025.600 2025.603 2025.605 2025.608 2025.611
#> [209] 2025.614 2025.616 2025.619 2025.622 2025.625 2025.627 2025.630 2025.633
#> [217] 2025.636 2025.638 2025.641 2025.644 2025.647 2025.649 2025.652 2025.655
#> [225] 2025.658 2025.660 2025.663 2025.666 2025.668 2025.668 2025.674 2025.677
#> [233] 2025.679 2025.682 2025.685 2025.688 2025.690 2025.693 2025.696 2025.699
#> [241] 2025.701 2025.704 2025.707 2025.710 2025.712 2025.715 2025.718 2025.721
#> [249] 2025.723 2025.726 2025.729 2025.732 2025.734 2025.737 2025.740 2025.742
#> [257] 2025.745 2025.748 2025.751 2025.751 2025.756 2025.759 2025.762 2025.764
#> [265] 2025.767 2025.770 2025.773 2025.775 2025.778 2025.781 2025.784 2025.786
#> [273] 2025.789 2025.792 2025.795 2025.797 2025.800 2025.803 2025.805 2025.808
#> [281] 2025.811 2025.814 2025.816 2025.819 2025.822 2025.825 2025.827 2025.830
#> [289] 2025.833 2025.836 2025.836 2025.841 2025.844 2025.847 2025.849 2025.852
#> [297] 2025.855 2025.858 2025.860 2025.863 2025.866 2025.868 2025.871 2025.874
#> [305] 2025.877 2025.879 2025.882 2025.885 2025.888 2025.890 2025.893 2025.896
#> [313] 2025.899 2025.901 2025.904 2025.907 2025.910 2025.912 2025.915 2025.918
#> [321] 2025.918 2025.923 2025.926 2025.929 2025.932 2025.934 2025.937 2025.940
#> [329] 2025.942 2025.945 2025.948 2025.951 2025.953 2025.956 2025.959 2025.962
#> [337] 2025.964 2025.967 2025.970 2025.973 2025.975 2025.978 2025.981 2025.984
#> [345] 2025.986 2025.989 2025.992 2025.995 2025.997 2026.000 2026.003 2026.005
#> [353] 2026.008 2026.011 2026.014 2026.016 2026.019 2026.022 2026.025 2026.027
#> [361] 2026.030 2026.033 2026.036 2026.038 2026.041 2026.044
#>