Get colors for plots

Gets colors used for WRTDS plots

gradcols(col_vec = NULL)

Arguments

col_vec

chr string of plot colors to use, typically passed to scale_colour_gradientn for shading. Any color palette from RColorBrewer can be used as a named input. Palettes from grDevices must be supplied as the returned string of colors for each palette.

Value

A character vector of colors in hexadecimal notation.

Details

This is a convenience function for retrieving a color palette that is used by most of the plotting functions. Palettes from RColorBrewer will use the maximum number of colors. The default palette is 'Spectral'.

See also

Examples

## defaults
gradcols()
#>  [1] "#9E0142" "#D53E4F" "#F46D43" "#FDAE61" "#FEE08B" "#FFFFBF" "#E6F598"
#>  [8] "#ABDDA4" "#66C2A5" "#3288BD" "#5E4FA2"

## another RColorBrewer palette
gradcols('Pastel2')
#> [1] "#B3E2CD" "#FDCDAC" "#CBD5E8" "#F4CAE4" "#E6F5C9" "#FFF2AE" "#F1E2CC"
#> [8] "#CCCCCC"

## a silly example
gradcols(rainbow(7))
#> [1] "#FF0000" "#FFDB00" "#49FF00" "#00FF92" "#0092FF" "#4900FF" "#FF00DB"