Create a grid of half-window widths to evaluate

Create a grid of all unique combinations of half-window widths to evaluate. The result can be passed to winsrch_grid.

createsrch(mos = c(seq(0.5, 1, by = 0.25), 2, 10), yrs = c(seq(5, 15,
  by = 3), 50), flo = c(seq(0.5, 1, by = 0.1), 5))

Arguments

mos

numeric vector of half-window widths for months, a value of one indicates twelve months

yrs

numeric vector of half-window widths for years, a value of one indicates one-year

flo

numeric vector of half-window widths for salinity or flow, a value of one indicates the full range of values (100 percent)

Value

A matrix with number of rows equal to the product of the lengths of each input vector, where each row is a unique combination for the selected half-window widths.

Details

The weighting function uses a tri-cube weighting scheme such that weights diminish with distance from the center of the window. For example, a value of one for the month window does not mean that all months are weighted equally even though the window covers an entire calendar year.

See also

Examples

createsrch()
#>       mos yrs flo
#> 1    0.50   5 0.5
#> 2    0.75   5 0.5
#> 3    1.00   5 0.5
#> 4    2.00   5 0.5
#> 5   10.00   5 0.5
#> 6    0.50   8 0.5
#> 7    0.75   8 0.5
#> 8    1.00   8 0.5
#> 9    2.00   8 0.5
#> 10  10.00   8 0.5
#> 11   0.50  11 0.5
#> 12   0.75  11 0.5
#> 13   1.00  11 0.5
#> 14   2.00  11 0.5
#> 15  10.00  11 0.5
#> 16   0.50  14 0.5
#> 17   0.75  14 0.5
#> 18   1.00  14 0.5
#> 19   2.00  14 0.5
#> 20  10.00  14 0.5
#> 21   0.50  50 0.5
#> 22   0.75  50 0.5
#> 23   1.00  50 0.5
#> 24   2.00  50 0.5
#> 25  10.00  50 0.5
#> 26   0.50   5 0.6
#> 27   0.75   5 0.6
#> 28   1.00   5 0.6
#> 29   2.00   5 0.6
#> 30  10.00   5 0.6
#> 31   0.50   8 0.6
#> 32   0.75   8 0.6
#> 33   1.00   8 0.6
#> 34   2.00   8 0.6
#> 35  10.00   8 0.6
#> 36   0.50  11 0.6
#> 37   0.75  11 0.6
#> 38   1.00  11 0.6
#> 39   2.00  11 0.6
#> 40  10.00  11 0.6
#> 41   0.50  14 0.6
#> 42   0.75  14 0.6
#> 43   1.00  14 0.6
#> 44   2.00  14 0.6
#> 45  10.00  14 0.6
#> 46   0.50  50 0.6
#> 47   0.75  50 0.6
#> 48   1.00  50 0.6
#> 49   2.00  50 0.6
#> 50  10.00  50 0.6
#> 51   0.50   5 0.7
#> 52   0.75   5 0.7
#> 53   1.00   5 0.7
#> 54   2.00   5 0.7
#> 55  10.00   5 0.7
#> 56   0.50   8 0.7
#> 57   0.75   8 0.7
#> 58   1.00   8 0.7
#> 59   2.00   8 0.7
#> 60  10.00   8 0.7
#> 61   0.50  11 0.7
#> 62   0.75  11 0.7
#> 63   1.00  11 0.7
#> 64   2.00  11 0.7
#> 65  10.00  11 0.7
#> 66   0.50  14 0.7
#> 67   0.75  14 0.7
#> 68   1.00  14 0.7
#> 69   2.00  14 0.7
#> 70  10.00  14 0.7
#> 71   0.50  50 0.7
#> 72   0.75  50 0.7
#> 73   1.00  50 0.7
#> 74   2.00  50 0.7
#> 75  10.00  50 0.7
#> 76   0.50   5 0.8
#> 77   0.75   5 0.8
#> 78   1.00   5 0.8
#> 79   2.00   5 0.8
#> 80  10.00   5 0.8
#> 81   0.50   8 0.8
#> 82   0.75   8 0.8
#> 83   1.00   8 0.8
#> 84   2.00   8 0.8
#> 85  10.00   8 0.8
#> 86   0.50  11 0.8
#> 87   0.75  11 0.8
#> 88   1.00  11 0.8
#> 89   2.00  11 0.8
#> 90  10.00  11 0.8
#> 91   0.50  14 0.8
#> 92   0.75  14 0.8
#> 93   1.00  14 0.8
#> 94   2.00  14 0.8
#> 95  10.00  14 0.8
#> 96   0.50  50 0.8
#> 97   0.75  50 0.8
#> 98   1.00  50 0.8
#> 99   2.00  50 0.8
#> 100 10.00  50 0.8
#> 101  0.50   5 0.9
#> 102  0.75   5 0.9
#> 103  1.00   5 0.9
#> 104  2.00   5 0.9
#> 105 10.00   5 0.9
#> 106  0.50   8 0.9
#> 107  0.75   8 0.9
#> 108  1.00   8 0.9
#> 109  2.00   8 0.9
#> 110 10.00   8 0.9
#> 111  0.50  11 0.9
#> 112  0.75  11 0.9
#> 113  1.00  11 0.9
#> 114  2.00  11 0.9
#> 115 10.00  11 0.9
#> 116  0.50  14 0.9
#> 117  0.75  14 0.9
#> 118  1.00  14 0.9
#> 119  2.00  14 0.9
#> 120 10.00  14 0.9
#> 121  0.50  50 0.9
#> 122  0.75  50 0.9
#> 123  1.00  50 0.9
#> 124  2.00  50 0.9
#> 125 10.00  50 0.9
#> 126  0.50   5 1.0
#> 127  0.75   5 1.0
#> 128  1.00   5 1.0
#> 129  2.00   5 1.0
#> 130 10.00   5 1.0
#> 131  0.50   8 1.0
#> 132  0.75   8 1.0
#> 133  1.00   8 1.0
#> 134  2.00   8 1.0
#> 135 10.00   8 1.0
#> 136  0.50  11 1.0
#> 137  0.75  11 1.0
#> 138  1.00  11 1.0
#> 139  2.00  11 1.0
#> 140 10.00  11 1.0
#> 141  0.50  14 1.0
#> 142  0.75  14 1.0
#> 143  1.00  14 1.0
#> 144  2.00  14 1.0
#> 145 10.00  14 1.0
#> 146  0.50  50 1.0
#> 147  0.75  50 1.0
#> 148  1.00  50 1.0
#> 149  2.00  50 1.0
#> 150 10.00  50 1.0
#> 151  0.50   5 5.0
#> 152  0.75   5 5.0
#> 153  1.00   5 5.0
#> 154  2.00   5 5.0
#> 155 10.00   5 5.0
#> 156  0.50   8 5.0
#> 157  0.75   8 5.0
#> 158  1.00   8 5.0
#> 159  2.00   8 5.0
#> 160 10.00   8 5.0
#> 161  0.50  11 5.0
#> 162  0.75  11 5.0
#> 163  1.00  11 5.0
#> 164  2.00  11 5.0
#> 165 10.00  11 5.0
#> 166  0.50  14 5.0
#> 167  0.75  14 5.0
#> 168  1.00  14 5.0
#> 169  2.00  14 5.0
#> 170 10.00  14 5.0
#> 171  0.50  50 5.0
#> 172  0.75  50 5.0
#> 173  1.00  50 5.0
#> 174  2.00  50 5.0
#> 175 10.00  50 5.0
createsrch(1, 1, 1)
#>   mos yrs flo
#> 1   1   1   1