Create decimal time from time vector

Create decimal time on an annual scale from an input time vector

dec_time(date_in)

# S3 method for Date
dec_time(date_in)

Arguments

date_in
input time vector, usually a date object

Value

A named list of four numeric vectors including day_num (decimal day on an annual scale), month (month of the year as integer), year, and dec_time (decimal time as sum of year and day_num)

Details

Function is used internally within the package.

Examples

dt <- Sys.Date()
dts <- seq.Date(dt - 365, dt, by = 'day')

dec_time(dts)

#> $day_num
#>   [1] 0.717808219 0.720547945 0.723287671 0.726027397 0.728767123 0.731506849
#>   [7] 0.734246575 0.736986301 0.739726027 0.742465753 0.745205479 0.747945205
#>  [13] 0.750684932 0.750684932 0.756164384 0.758904110 0.761643836 0.764383562
#>  [19] 0.767123288 0.769863014 0.772602740 0.775342466 0.778082192 0.780821918
#>  [25] 0.783561644 0.786301370 0.789041096 0.791780822 0.794520548 0.797260274
#>  [31] 0.800000000 0.802739726 0.805479452 0.808219178 0.810958904 0.813698630
#>  [37] 0.816438356 0.819178082 0.821917808 0.824657534 0.827397260 0.830136986
#>  [43] 0.832876712 0.835616438 0.835616438 0.841095890 0.843835616 0.846575342
#>  [49] 0.849315068 0.852054795 0.854794521 0.857534247 0.860273973 0.863013699
#>  [55] 0.865753425 0.868493151 0.871232877 0.873972603 0.876712329 0.879452055
#>  [61] 0.882191781 0.884931507 0.887671233 0.890410959 0.893150685 0.895890411
#>  [67] 0.898630137 0.901369863 0.904109589 0.906849315 0.909589041 0.912328767
#>  [73] 0.915068493 0.917808219 0.917808219 0.923287671 0.926027397 0.928767123
#>  [79] 0.931506849 0.934246575 0.936986301 0.939726027 0.942465753 0.945205479
#>  [85] 0.947945205 0.950684932 0.953424658 0.956164384 0.958904110 0.961643836
#>  [91] 0.964383562 0.967123288 0.969863014 0.972602740 0.975342466 0.978082192
#>  [97] 0.980821918 0.983561644 0.986301370 0.989041096 0.991780822 0.994520548
#> [103] 0.997260274 1.000000000 1.002739726 0.005479452 0.008219178 0.010958904
#> [109] 0.013698630 0.016438356 0.019178082 0.021917808 0.024657534 0.027397260
#> [115] 0.030136986 0.032876712 0.035616438 0.038356164 0.041095890 0.043835616
#> [121] 0.046575342 0.049315068 0.052054795 0.054794521 0.057534247 0.060273973
#> [127] 0.063013699 0.065753425 0.068493151 0.071232877 0.073972603 0.076712329
#> [133] 0.079452055 0.082191781 0.084931507 0.087671233 0.090410959 0.093150685
#> [139] 0.095890411 0.098630137 0.101369863 0.104109589 0.106849315 0.109589041
#> [145] 0.112328767 0.115068493 0.117808219 0.120547945 0.123287671 0.126027397
#> [151] 0.128767123 0.131506849 0.134246575 0.136986301 0.139726027 0.142465753
#> [157] 0.145205479 0.147945205 0.150684932 0.153424658 0.156164384 0.158904110
#> [163] 0.161643836 0.164383562 0.164383562 0.169863014 0.172602740 0.175342466
#> [169] 0.178082192 0.180821918 0.183561644 0.186301370 0.189041096 0.191780822
#> [175] 0.194520548 0.197260274 0.200000000 0.202739726 0.205479452 0.208219178
#> [181] 0.210958904 0.213698630 0.216438356 0.219178082 0.221917808 0.224657534
#> [187] 0.227397260 0.230136986 0.232876712 0.235616438 0.238356164 0.241095890
#> [193] 0.243835616 0.246575342 0.249315068 0.249315068 0.254794521 0.257534247
#> [199] 0.260273973 0.263013699 0.265753425 0.268493151 0.271232877 0.273972603
#> [205] 0.276712329 0.279452055 0.282191781 0.284931507 0.287671233 0.290410959
#> [211] 0.293150685 0.295890411 0.298630137 0.301369863 0.304109589 0.306849315
#> [217] 0.309589041 0.312328767 0.315068493 0.317808219 0.320547945 0.323287671
#> [223] 0.326027397 0.328767123 0.331506849 0.331506849 0.336986301 0.339726027
#> [229] 0.342465753 0.345205479 0.347945205 0.350684932 0.353424658 0.356164384
#> [235] 0.358904110 0.361643836 0.364383562 0.367123288 0.369863014 0.372602740
#> [241] 0.375342466 0.378082192 0.380821918 0.383561644 0.386301370 0.389041096
#> [247] 0.391780822 0.394520548 0.397260274 0.400000000 0.402739726 0.405479452
#> [253] 0.408219178 0.410958904 0.413698630 0.416438356 0.416438356 0.421917808
#> [259] 0.424657534 0.427397260 0.430136986 0.432876712 0.435616438 0.438356164
#> [265] 0.441095890 0.443835616 0.446575342 0.449315068 0.452054795 0.454794521
#> [271] 0.457534247 0.460273973 0.463013699 0.465753425 0.468493151 0.471232877
#> [277] 0.473972603 0.476712329 0.479452055 0.482191781 0.484931507 0.487671233
#> [283] 0.490410959 0.493150685 0.495890411 0.498630137 0.498630137 0.504109589
#> [289] 0.506849315 0.509589041 0.512328767 0.515068493 0.517808219 0.520547945
#> [295] 0.523287671 0.526027397 0.528767123 0.531506849 0.534246575 0.536986301
#> [301] 0.539726027 0.542465753 0.545205479 0.547945205 0.550684932 0.553424658
#> [307] 0.556164384 0.558904110 0.561643836 0.564383562 0.567123288 0.569863014
#> [313] 0.572602740 0.575342466 0.578082192 0.580821918 0.583561644 0.583561644
#> [319] 0.589041096 0.591780822 0.594520548 0.597260274 0.600000000 0.602739726
#> [325] 0.605479452 0.608219178 0.610958904 0.613698630 0.616438356 0.619178082
#> [331] 0.621917808 0.624657534 0.627397260 0.630136986 0.632876712 0.635616438
#> [337] 0.638356164 0.641095890 0.643835616 0.646575342 0.649315068 0.652054795
#> [343] 0.654794521 0.657534247 0.660273973 0.663013699 0.665753425 0.668493151
#> [349] 0.668493151 0.673972603 0.676712329 0.679452055 0.682191781 0.684931507
#> [355] 0.687671233 0.690410959 0.693150685 0.695890411 0.698630137 0.701369863
#> [361] 0.704109589 0.706849315 0.709589041 0.712328767 0.715068493 0.717808219
#> 
#> $month
#>   [1]  9  9  9  9  9  9  9  9  9  9  9  9  9 10 10 10 10 10 10 10 10 10 10 10 10
#>  [26] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 11 11 11 11 11 11
#>  [51] 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 12
#>  [76] 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12
#> [101] 12 12 12 12 12  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> [126]  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2
#> [151]  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  3  3
#> [176]  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  4  4  4  4
#> [201]  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4
#> [226]  4  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5
#> [251]  5  5  5  5  5  5  5  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#> [276]  6  6  6  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7  7  7  7  7  7  7
#> [301]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  8  8  8  8  8  8  8
#> [326]  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  9
#> [351]  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9
#> 
#> $year
#>   [1] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [16] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [31] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [46] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [61] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [76] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [91] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [106] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [121] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [136] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [151] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [166] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [181] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [196] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [211] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [226] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [241] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [256] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [271] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [286] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [301] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [316] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [331] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [346] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020
#> [361] 2020 2020 2020 2020 2020 2020
#> 
#> $dec_time
#>   [1] 2019.718 2019.721 2019.723 2019.726 2019.729 2019.732 2019.734 2019.737
#>   [9] 2019.740 2019.742 2019.745 2019.748 2019.751 2019.751 2019.756 2019.759
#>  [17] 2019.762 2019.764 2019.767 2019.770 2019.773 2019.775 2019.778 2019.781
#>  [25] 2019.784 2019.786 2019.789 2019.792 2019.795 2019.797 2019.800 2019.803
#>  [33] 2019.805 2019.808 2019.811 2019.814 2019.816 2019.819 2019.822 2019.825
#>  [41] 2019.827 2019.830 2019.833 2019.836 2019.836 2019.841 2019.844 2019.847
#>  [49] 2019.849 2019.852 2019.855 2019.858 2019.860 2019.863 2019.866 2019.868
#>  [57] 2019.871 2019.874 2019.877 2019.879 2019.882 2019.885 2019.888 2019.890
#>  [65] 2019.893 2019.896 2019.899 2019.901 2019.904 2019.907 2019.910 2019.912
#>  [73] 2019.915 2019.918 2019.918 2019.923 2019.926 2019.929 2019.932 2019.934
#>  [81] 2019.937 2019.940 2019.942 2019.945 2019.948 2019.951 2019.953 2019.956
#>  [89] 2019.959 2019.962 2019.964 2019.967 2019.970 2019.973 2019.975 2019.978
#>  [97] 2019.981 2019.984 2019.986 2019.989 2019.992 2019.995 2019.997 2020.000
#> [105] 2020.003 2020.005 2020.008 2020.011 2020.014 2020.016 2020.019 2020.022
#> [113] 2020.025 2020.027 2020.030 2020.033 2020.036 2020.038 2020.041 2020.044
#> [121] 2020.047 2020.049 2020.052 2020.055 2020.058 2020.060 2020.063 2020.066
#> [129] 2020.068 2020.071 2020.074 2020.077 2020.079 2020.082 2020.085 2020.088
#> [137] 2020.090 2020.093 2020.096 2020.099 2020.101 2020.104 2020.107 2020.110
#> [145] 2020.112 2020.115 2020.118 2020.121 2020.123 2020.126 2020.129 2020.132
#> [153] 2020.134 2020.137 2020.140 2020.142 2020.145 2020.148 2020.151 2020.153
#> [161] 2020.156 2020.159 2020.162 2020.164 2020.164 2020.170 2020.173 2020.175
#> [169] 2020.178 2020.181 2020.184 2020.186 2020.189 2020.192 2020.195 2020.197
#> [177] 2020.200 2020.203 2020.205 2020.208 2020.211 2020.214 2020.216 2020.219
#> [185] 2020.222 2020.225 2020.227 2020.230 2020.233 2020.236 2020.238 2020.241
#> [193] 2020.244 2020.247 2020.249 2020.249 2020.255 2020.258 2020.260 2020.263
#> [201] 2020.266 2020.268 2020.271 2020.274 2020.277 2020.279 2020.282 2020.285
#> [209] 2020.288 2020.290 2020.293 2020.296 2020.299 2020.301 2020.304 2020.307
#> [217] 2020.310 2020.312 2020.315 2020.318 2020.321 2020.323 2020.326 2020.329
#> [225] 2020.332 2020.332 2020.337 2020.340 2020.342 2020.345 2020.348 2020.351
#> [233] 2020.353 2020.356 2020.359 2020.362 2020.364 2020.367 2020.370 2020.373
#> [241] 2020.375 2020.378 2020.381 2020.384 2020.386 2020.389 2020.392 2020.395
#> [249] 2020.397 2020.400 2020.403 2020.405 2020.408 2020.411 2020.414 2020.416
#> [257] 2020.416 2020.422 2020.425 2020.427 2020.430 2020.433 2020.436 2020.438
#> [265] 2020.441 2020.444 2020.447 2020.449 2020.452 2020.455 2020.458 2020.460
#> [273] 2020.463 2020.466 2020.468 2020.471 2020.474 2020.477 2020.479 2020.482
#> [281] 2020.485 2020.488 2020.490 2020.493 2020.496 2020.499 2020.499 2020.504
#> [289] 2020.507 2020.510 2020.512 2020.515 2020.518 2020.521 2020.523 2020.526
#> [297] 2020.529 2020.532 2020.534 2020.537 2020.540 2020.542 2020.545 2020.548
#> [305] 2020.551 2020.553 2020.556 2020.559 2020.562 2020.564 2020.567 2020.570
#> [313] 2020.573 2020.575 2020.578 2020.581 2020.584 2020.584 2020.589 2020.592
#> [321] 2020.595 2020.597 2020.600 2020.603 2020.605 2020.608 2020.611 2020.614
#> [329] 2020.616 2020.619 2020.622 2020.625 2020.627 2020.630 2020.633 2020.636
#> [337] 2020.638 2020.641 2020.644 2020.647 2020.649 2020.652 2020.655 2020.658
#> [345] 2020.660 2020.663 2020.666 2020.668 2020.668 2020.674 2020.677 2020.679
#> [353] 2020.682 2020.685 2020.688 2020.690 2020.693 2020.696 2020.699 2020.701
#> [361] 2020.704 2020.707 2020.710 2020.712 2020.715 2020.718
#>