Create decimal time from time vector

Create decimal time on an annual scale from an input time vector

dec_time(date_in)

# S3 method for Date
dec_time(date_in)

Arguments

date_in
input time vector, usually a date object

Value

A named list of four numeric vectors including day_num (decimal day on an annual scale), month (month of the year as integer), year, and dec_time (decimal time as sum of year and day_num)

Details

Function is used internally within the package.

Examples

dt <- Sys.Date()
dts <- seq.Date(dt - 365, dt, by = 'day')

dec_time(dts)

#> $day_num
#>   [1] 0.884931507 0.887671233 0.890410959 0.893150685 0.895890411 0.898630137
#>   [7] 0.901369863 0.904109589 0.906849315 0.909589041 0.912328767 0.915068493
#>  [13] 0.917808219 0.917808219 0.923287671 0.926027397 0.928767123 0.931506849
#>  [19] 0.934246575 0.936986301 0.939726027 0.942465753 0.945205479 0.947945205
#>  [25] 0.950684932 0.953424658 0.956164384 0.958904110 0.961643836 0.964383562
#>  [31] 0.967123288 0.969863014 0.972602740 0.975342466 0.978082192 0.980821918
#>  [37] 0.983561644 0.986301370 0.989041096 0.991780822 0.994520548 0.997260274
#>  [43] 1.000000000 1.002739726 0.005479452 0.008219178 0.010958904 0.013698630
#>  [49] 0.016438356 0.019178082 0.021917808 0.024657534 0.027397260 0.030136986
#>  [55] 0.032876712 0.035616438 0.038356164 0.041095890 0.043835616 0.046575342
#>  [61] 0.049315068 0.052054795 0.054794521 0.057534247 0.060273973 0.063013699
#>  [67] 0.065753425 0.068493151 0.071232877 0.073972603 0.076712329 0.079452055
#>  [73] 0.082191781 0.084931507 0.087671233 0.090410959 0.093150685 0.095890411
#>  [79] 0.098630137 0.101369863 0.104109589 0.106849315 0.109589041 0.112328767
#>  [85] 0.115068493 0.117808219 0.120547945 0.123287671 0.126027397 0.128767123
#>  [91] 0.131506849 0.134246575 0.136986301 0.139726027 0.142465753 0.145205479
#>  [97] 0.147945205 0.150684932 0.153424658 0.156164384 0.158904110 0.161643836
#> [103] 0.164383562 0.164383562 0.169863014 0.172602740 0.175342466 0.178082192
#> [109] 0.180821918 0.183561644 0.186301370 0.189041096 0.191780822 0.194520548
#> [115] 0.197260274 0.200000000 0.202739726 0.205479452 0.208219178 0.210958904
#> [121] 0.213698630 0.216438356 0.219178082 0.221917808 0.224657534 0.227397260
#> [127] 0.230136986 0.232876712 0.235616438 0.238356164 0.241095890 0.243835616
#> [133] 0.246575342 0.249315068 0.249315068 0.254794521 0.257534247 0.260273973
#> [139] 0.263013699 0.265753425 0.268493151 0.271232877 0.273972603 0.276712329
#> [145] 0.279452055 0.282191781 0.284931507 0.287671233 0.290410959 0.293150685
#> [151] 0.295890411 0.298630137 0.301369863 0.304109589 0.306849315 0.309589041
#> [157] 0.312328767 0.315068493 0.317808219 0.320547945 0.323287671 0.326027397
#> [163] 0.328767123 0.331506849 0.331506849 0.336986301 0.339726027 0.342465753
#> [169] 0.345205479 0.347945205 0.350684932 0.353424658 0.356164384 0.358904110
#> [175] 0.361643836 0.364383562 0.367123288 0.369863014 0.372602740 0.375342466
#> [181] 0.378082192 0.380821918 0.383561644 0.386301370 0.389041096 0.391780822
#> [187] 0.394520548 0.397260274 0.400000000 0.402739726 0.405479452 0.408219178
#> [193] 0.410958904 0.413698630 0.416438356 0.416438356 0.421917808 0.424657534
#> [199] 0.427397260 0.430136986 0.432876712 0.435616438 0.438356164 0.441095890
#> [205] 0.443835616 0.446575342 0.449315068 0.452054795 0.454794521 0.457534247
#> [211] 0.460273973 0.463013699 0.465753425 0.468493151 0.471232877 0.473972603
#> [217] 0.476712329 0.479452055 0.482191781 0.484931507 0.487671233 0.490410959
#> [223] 0.493150685 0.495890411 0.498630137 0.498630137 0.504109589 0.506849315
#> [229] 0.509589041 0.512328767 0.515068493 0.517808219 0.520547945 0.523287671
#> [235] 0.526027397 0.528767123 0.531506849 0.534246575 0.536986301 0.539726027
#> [241] 0.542465753 0.545205479 0.547945205 0.550684932 0.553424658 0.556164384
#> [247] 0.558904110 0.561643836 0.564383562 0.567123288 0.569863014 0.572602740
#> [253] 0.575342466 0.578082192 0.580821918 0.583561644 0.583561644 0.589041096
#> [259] 0.591780822 0.594520548 0.597260274 0.600000000 0.602739726 0.605479452
#> [265] 0.608219178 0.610958904 0.613698630 0.616438356 0.619178082 0.621917808
#> [271] 0.624657534 0.627397260 0.630136986 0.632876712 0.635616438 0.638356164
#> [277] 0.641095890 0.643835616 0.646575342 0.649315068 0.652054795 0.654794521
#> [283] 0.657534247 0.660273973 0.663013699 0.665753425 0.668493151 0.668493151
#> [289] 0.673972603 0.676712329 0.679452055 0.682191781 0.684931507 0.687671233
#> [295] 0.690410959 0.693150685 0.695890411 0.698630137 0.701369863 0.704109589
#> [301] 0.706849315 0.709589041 0.712328767 0.715068493 0.717808219 0.720547945
#> [307] 0.723287671 0.726027397 0.728767123 0.731506849 0.734246575 0.736986301
#> [313] 0.739726027 0.742465753 0.745205479 0.747945205 0.750684932 0.750684932
#> [319] 0.756164384 0.758904110 0.761643836 0.764383562 0.767123288 0.769863014
#> [325] 0.772602740 0.775342466 0.778082192 0.780821918 0.783561644 0.786301370
#> [331] 0.789041096 0.791780822 0.794520548 0.797260274 0.800000000 0.802739726
#> [337] 0.805479452 0.808219178 0.810958904 0.813698630 0.816438356 0.819178082
#> [343] 0.821917808 0.824657534 0.827397260 0.830136986 0.832876712 0.835616438
#> [349] 0.835616438 0.841095890 0.843835616 0.846575342 0.849315068 0.852054795
#> [355] 0.854794521 0.857534247 0.860273973 0.863013699 0.865753425 0.868493151
#> [361] 0.871232877 0.873972603 0.876712329 0.879452055 0.882191781 0.884931507
#> 
#> $month
#>   [1] 11 11 11 11 11 11 11 11 11 11 11 11 11 12 12 12 12 12 12 12 12 12 12 12 12
#>  [26] 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12  1  1  1  1  1  1
#>  [51]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>  [76]  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2
#> [101]  2  2  2  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3
#> [126]  3  3  3  3  3  3  3  3  3  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4
#> [151]  4  4  4  4  4  4  4  4  4  4  4  4  4  4  5  5  5  5  5  5  5  5  5  5  5
#> [176]  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  6  6  6  6  6
#> [201]  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#> [226]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#> [251]  7  7  7  7  7  7  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8
#> [276]  8  8  8  8  8  8  8  8  8  8  8  8  9  9  9  9  9  9  9  9  9  9  9  9  9
#> [301]  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9 10 10 10 10 10 10 10 10
#> [326] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 11 11
#> [351] 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11
#> 
#> $year
#>   [1] 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018
#>  [16] 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018
#>  [31] 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 2019
#>  [46] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [61] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [76] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#>  [91] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [106] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [121] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [136] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [151] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [166] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [181] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [196] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [211] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [226] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [241] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [256] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [271] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [286] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [301] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [316] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [331] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [346] 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019 2019
#> [361] 2019 2019 2019 2019 2019 2019
#> 
#> $dec_time
#>   [1] 2018.885 2018.888 2018.890 2018.893 2018.896 2018.899 2018.901 2018.904
#>   [9] 2018.907 2018.910 2018.912 2018.915 2018.918 2018.918 2018.923 2018.926
#>  [17] 2018.929 2018.932 2018.934 2018.937 2018.940 2018.942 2018.945 2018.948
#>  [25] 2018.951 2018.953 2018.956 2018.959 2018.962 2018.964 2018.967 2018.970
#>  [33] 2018.973 2018.975 2018.978 2018.981 2018.984 2018.986 2018.989 2018.992
#>  [41] 2018.995 2018.997 2019.000 2019.003 2019.005 2019.008 2019.011 2019.014
#>  [49] 2019.016 2019.019 2019.022 2019.025 2019.027 2019.030 2019.033 2019.036
#>  [57] 2019.038 2019.041 2019.044 2019.047 2019.049 2019.052 2019.055 2019.058
#>  [65] 2019.060 2019.063 2019.066 2019.068 2019.071 2019.074 2019.077 2019.079
#>  [73] 2019.082 2019.085 2019.088 2019.090 2019.093 2019.096 2019.099 2019.101
#>  [81] 2019.104 2019.107 2019.110 2019.112 2019.115 2019.118 2019.121 2019.123
#>  [89] 2019.126 2019.129 2019.132 2019.134 2019.137 2019.140 2019.142 2019.145
#>  [97] 2019.148 2019.151 2019.153 2019.156 2019.159 2019.162 2019.164 2019.164
#> [105] 2019.170 2019.173 2019.175 2019.178 2019.181 2019.184 2019.186 2019.189
#> [113] 2019.192 2019.195 2019.197 2019.200 2019.203 2019.205 2019.208 2019.211
#> [121] 2019.214 2019.216 2019.219 2019.222 2019.225 2019.227 2019.230 2019.233
#> [129] 2019.236 2019.238 2019.241 2019.244 2019.247 2019.249 2019.249 2019.255
#> [137] 2019.258 2019.260 2019.263 2019.266 2019.268 2019.271 2019.274 2019.277
#> [145] 2019.279 2019.282 2019.285 2019.288 2019.290 2019.293 2019.296 2019.299
#> [153] 2019.301 2019.304 2019.307 2019.310 2019.312 2019.315 2019.318 2019.321
#> [161] 2019.323 2019.326 2019.329 2019.332 2019.332 2019.337 2019.340 2019.342
#> [169] 2019.345 2019.348 2019.351 2019.353 2019.356 2019.359 2019.362 2019.364
#> [177] 2019.367 2019.370 2019.373 2019.375 2019.378 2019.381 2019.384 2019.386
#> [185] 2019.389 2019.392 2019.395 2019.397 2019.400 2019.403 2019.405 2019.408
#> [193] 2019.411 2019.414 2019.416 2019.416 2019.422 2019.425 2019.427 2019.430
#> [201] 2019.433 2019.436 2019.438 2019.441 2019.444 2019.447 2019.449 2019.452
#> [209] 2019.455 2019.458 2019.460 2019.463 2019.466 2019.468 2019.471 2019.474
#> [217] 2019.477 2019.479 2019.482 2019.485 2019.488 2019.490 2019.493 2019.496
#> [225] 2019.499 2019.499 2019.504 2019.507 2019.510 2019.512 2019.515 2019.518
#> [233] 2019.521 2019.523 2019.526 2019.529 2019.532 2019.534 2019.537 2019.540
#> [241] 2019.542 2019.545 2019.548 2019.551 2019.553 2019.556 2019.559 2019.562
#> [249] 2019.564 2019.567 2019.570 2019.573 2019.575 2019.578 2019.581 2019.584
#> [257] 2019.584 2019.589 2019.592 2019.595 2019.597 2019.600 2019.603 2019.605
#> [265] 2019.608 2019.611 2019.614 2019.616 2019.619 2019.622 2019.625 2019.627
#> [273] 2019.630 2019.633 2019.636 2019.638 2019.641 2019.644 2019.647 2019.649
#> [281] 2019.652 2019.655 2019.658 2019.660 2019.663 2019.666 2019.668 2019.668
#> [289] 2019.674 2019.677 2019.679 2019.682 2019.685 2019.688 2019.690 2019.693
#> [297] 2019.696 2019.699 2019.701 2019.704 2019.707 2019.710 2019.712 2019.715
#> [305] 2019.718 2019.721 2019.723 2019.726 2019.729 2019.732 2019.734 2019.737
#> [313] 2019.740 2019.742 2019.745 2019.748 2019.751 2019.751 2019.756 2019.759
#> [321] 2019.762 2019.764 2019.767 2019.770 2019.773 2019.775 2019.778 2019.781
#> [329] 2019.784 2019.786 2019.789 2019.792 2019.795 2019.797 2019.800 2019.803
#> [337] 2019.805 2019.808 2019.811 2019.814 2019.816 2019.819 2019.822 2019.825
#> [345] 2019.827 2019.830 2019.833 2019.836 2019.836 2019.841 2019.844 2019.847
#> [353] 2019.849 2019.852 2019.855 2019.858 2019.860 2019.863 2019.866 2019.868
#> [361] 2019.871 2019.874 2019.877 2019.879 2019.882 2019.885
#>