Create decimal time from time vector

Create decimal time on an annual scale from an input time vector

dec_time(date_in)

# S3 method for Date
dec_time(date_in)

Arguments

date_in

input time vector, usually a date object

Value

A named list of four numeric vectors including day_num (decimal day on an annual scale), month (month of the year as integer), year, and dec_time (decimal time as sum of year and day_num)

Details

Function is used internally within the package.

Examples

dt <- Sys.Date()
dts <- seq.Date(dt - 365, dt, by = 'day') 

dec_time(dts)
#> $day_num
#>   [1] 0.471232877 0.473972603 0.476712329 0.479452055 0.482191781 0.484931507
#>   [7] 0.487671233 0.490410959 0.493150685 0.495890411 0.498630137 0.498630137
#>  [13] 0.504109589 0.506849315 0.509589041 0.512328767 0.515068493 0.517808219
#>  [19] 0.520547945 0.523287671 0.526027397 0.528767123 0.531506849 0.534246575
#>  [25] 0.536986301 0.539726027 0.542465753 0.545205479 0.547945205 0.550684932
#>  [31] 0.553424658 0.556164384 0.558904110 0.561643836 0.564383562 0.567123288
#>  [37] 0.569863014 0.572602740 0.575342466 0.578082192 0.580821918 0.583561644
#>  [43] 0.583561644 0.589041096 0.591780822 0.594520548 0.597260274 0.600000000
#>  [49] 0.602739726 0.605479452 0.608219178 0.610958904 0.613698630 0.616438356
#>  [55] 0.619178082 0.621917808 0.624657534 0.627397260 0.630136986 0.632876712
#>  [61] 0.635616438 0.638356164 0.641095890 0.643835616 0.646575342 0.649315068
#>  [67] 0.652054795 0.654794521 0.657534247 0.660273973 0.663013699 0.665753425
#>  [73] 0.668493151 0.668493151 0.673972603 0.676712329 0.679452055 0.682191781
#>  [79] 0.684931507 0.687671233 0.690410959 0.693150685 0.695890411 0.698630137
#>  [85] 0.701369863 0.704109589 0.706849315 0.709589041 0.712328767 0.715068493
#>  [91] 0.717808219 0.720547945 0.723287671 0.726027397 0.728767123 0.731506849
#>  [97] 0.734246575 0.736986301 0.739726027 0.742465753 0.745205479 0.747945205
#> [103] 0.750684932 0.750684932 0.756164384 0.758904110 0.761643836 0.764383562
#> [109] 0.767123288 0.769863014 0.772602740 0.775342466 0.778082192 0.780821918
#> [115] 0.783561644 0.786301370 0.789041096 0.791780822 0.794520548 0.797260274
#> [121] 0.800000000 0.802739726 0.805479452 0.808219178 0.810958904 0.813698630
#> [127] 0.816438356 0.819178082 0.821917808 0.824657534 0.827397260 0.830136986
#> [133] 0.832876712 0.835616438 0.835616438 0.841095890 0.843835616 0.846575342
#> [139] 0.849315068 0.852054795 0.854794521 0.857534247 0.860273973 0.863013699
#> [145] 0.865753425 0.868493151 0.871232877 0.873972603 0.876712329 0.879452055
#> [151] 0.882191781 0.884931507 0.887671233 0.890410959 0.893150685 0.895890411
#> [157] 0.898630137 0.901369863 0.904109589 0.906849315 0.909589041 0.912328767
#> [163] 0.915068493 0.917808219 0.917808219 0.923287671 0.926027397 0.928767123
#> [169] 0.931506849 0.934246575 0.936986301 0.939726027 0.942465753 0.945205479
#> [175] 0.947945205 0.950684932 0.953424658 0.956164384 0.958904110 0.961643836
#> [181] 0.964383562 0.967123288 0.969863014 0.972602740 0.975342466 0.978082192
#> [187] 0.980821918 0.983561644 0.986301370 0.989041096 0.991780822 0.994520548
#> [193] 0.997260274 1.000000000 1.002739726 0.005479452 0.008219178 0.010958904
#> [199] 0.013698630 0.016438356 0.019178082 0.021917808 0.024657534 0.027397260
#> [205] 0.030136986 0.032876712 0.035616438 0.038356164 0.041095890 0.043835616
#> [211] 0.046575342 0.049315068 0.052054795 0.054794521 0.057534247 0.060273973
#> [217] 0.063013699 0.065753425 0.068493151 0.071232877 0.073972603 0.076712329
#> [223] 0.079452055 0.082191781 0.084931507 0.087671233 0.090410959 0.093150685
#> [229] 0.095890411 0.098630137 0.101369863 0.104109589 0.106849315 0.109589041
#> [235] 0.112328767 0.115068493 0.117808219 0.120547945 0.123287671 0.126027397
#> [241] 0.128767123 0.131506849 0.134246575 0.136986301 0.139726027 0.142465753
#> [247] 0.145205479 0.147945205 0.150684932 0.153424658 0.156164384 0.158904110
#> [253] 0.161643836 0.164383562 0.164383562 0.169863014 0.172602740 0.175342466
#> [259] 0.178082192 0.180821918 0.183561644 0.186301370 0.189041096 0.191780822
#> [265] 0.194520548 0.197260274 0.200000000 0.202739726 0.205479452 0.208219178
#> [271] 0.210958904 0.213698630 0.216438356 0.219178082 0.221917808 0.224657534
#> [277] 0.227397260 0.230136986 0.232876712 0.235616438 0.238356164 0.241095890
#> [283] 0.243835616 0.246575342 0.249315068 0.249315068 0.254794521 0.257534247
#> [289] 0.260273973 0.263013699 0.265753425 0.268493151 0.271232877 0.273972603
#> [295] 0.276712329 0.279452055 0.282191781 0.284931507 0.287671233 0.290410959
#> [301] 0.293150685 0.295890411 0.298630137 0.301369863 0.304109589 0.306849315
#> [307] 0.309589041 0.312328767 0.315068493 0.317808219 0.320547945 0.323287671
#> [313] 0.326027397 0.328767123 0.331506849 0.331506849 0.336986301 0.339726027
#> [319] 0.342465753 0.345205479 0.347945205 0.350684932 0.353424658 0.356164384
#> [325] 0.358904110 0.361643836 0.364383562 0.367123288 0.369863014 0.372602740
#> [331] 0.375342466 0.378082192 0.380821918 0.383561644 0.386301370 0.389041096
#> [337] 0.391780822 0.394520548 0.397260274 0.400000000 0.402739726 0.405479452
#> [343] 0.408219178 0.410958904 0.413698630 0.416438356 0.416438356 0.421917808
#> [349] 0.424657534 0.427397260 0.430136986 0.432876712 0.435616438 0.438356164
#> [355] 0.441095890 0.443835616 0.446575342 0.449315068 0.452054795 0.454794521
#> [361] 0.457534247 0.460273973 0.463013699 0.465753425 0.468493151 0.471232877
#> 
#> $month
#>   [1]  6  6  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#>  [26]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  8  8  8  8  8  8  8  8
#>  [51]  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  9  9
#>  [76]  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9
#> [101]  9  9  9 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#> [126] 10 10 10 10 10 10 10 10 10 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11
#> [151] 11 11 11 11 11 11 11 11 11 11 11 11 11 11 12 12 12 12 12 12 12 12 12 12 12
#> [176] 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12  1  1  1  1  1
#> [201]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> [226]  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2
#> [251]  2  2  2  2  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3
#> [276]  3  3  3  3  3  3  3  3  3  3  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4
#> [301]  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  5  5  5  5  5  5  5  5  5  5
#> [326]  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  6  6  6  6
#> [351]  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#> 
#> $year
#>   [1] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#>  [16] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#>  [31] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#>  [46] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#>  [61] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#>  [76] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#>  [91] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#> [106] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#> [121] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#> [136] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#> [151] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#> [166] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#> [181] 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021 2021
#> [196] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [211] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [226] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [241] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [256] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [271] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [286] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [301] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [316] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [331] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [346] 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022 2022
#> [361] 2022 2022 2022 2022 2022 2022
#> 
#> $dec_time
#>   [1] 2021.471 2021.474 2021.477 2021.479 2021.482 2021.485 2021.488 2021.490
#>   [9] 2021.493 2021.496 2021.499 2021.499 2021.504 2021.507 2021.510 2021.512
#>  [17] 2021.515 2021.518 2021.521 2021.523 2021.526 2021.529 2021.532 2021.534
#>  [25] 2021.537 2021.540 2021.542 2021.545 2021.548 2021.551 2021.553 2021.556
#>  [33] 2021.559 2021.562 2021.564 2021.567 2021.570 2021.573 2021.575 2021.578
#>  [41] 2021.581 2021.584 2021.584 2021.589 2021.592 2021.595 2021.597 2021.600
#>  [49] 2021.603 2021.605 2021.608 2021.611 2021.614 2021.616 2021.619 2021.622
#>  [57] 2021.625 2021.627 2021.630 2021.633 2021.636 2021.638 2021.641 2021.644
#>  [65] 2021.647 2021.649 2021.652 2021.655 2021.658 2021.660 2021.663 2021.666
#>  [73] 2021.668 2021.668 2021.674 2021.677 2021.679 2021.682 2021.685 2021.688
#>  [81] 2021.690 2021.693 2021.696 2021.699 2021.701 2021.704 2021.707 2021.710
#>  [89] 2021.712 2021.715 2021.718 2021.721 2021.723 2021.726 2021.729 2021.732
#>  [97] 2021.734 2021.737 2021.740 2021.742 2021.745 2021.748 2021.751 2021.751
#> [105] 2021.756 2021.759 2021.762 2021.764 2021.767 2021.770 2021.773 2021.775
#> [113] 2021.778 2021.781 2021.784 2021.786 2021.789 2021.792 2021.795 2021.797
#> [121] 2021.800 2021.803 2021.805 2021.808 2021.811 2021.814 2021.816 2021.819
#> [129] 2021.822 2021.825 2021.827 2021.830 2021.833 2021.836 2021.836 2021.841
#> [137] 2021.844 2021.847 2021.849 2021.852 2021.855 2021.858 2021.860 2021.863
#> [145] 2021.866 2021.868 2021.871 2021.874 2021.877 2021.879 2021.882 2021.885
#> [153] 2021.888 2021.890 2021.893 2021.896 2021.899 2021.901 2021.904 2021.907
#> [161] 2021.910 2021.912 2021.915 2021.918 2021.918 2021.923 2021.926 2021.929
#> [169] 2021.932 2021.934 2021.937 2021.940 2021.942 2021.945 2021.948 2021.951
#> [177] 2021.953 2021.956 2021.959 2021.962 2021.964 2021.967 2021.970 2021.973
#> [185] 2021.975 2021.978 2021.981 2021.984 2021.986 2021.989 2021.992 2021.995
#> [193] 2021.997 2022.000 2022.003 2022.005 2022.008 2022.011 2022.014 2022.016
#> [201] 2022.019 2022.022 2022.025 2022.027 2022.030 2022.033 2022.036 2022.038
#> [209] 2022.041 2022.044 2022.047 2022.049 2022.052 2022.055 2022.058 2022.060
#> [217] 2022.063 2022.066 2022.068 2022.071 2022.074 2022.077 2022.079 2022.082
#> [225] 2022.085 2022.088 2022.090 2022.093 2022.096 2022.099 2022.101 2022.104
#> [233] 2022.107 2022.110 2022.112 2022.115 2022.118 2022.121 2022.123 2022.126
#> [241] 2022.129 2022.132 2022.134 2022.137 2022.140 2022.142 2022.145 2022.148
#> [249] 2022.151 2022.153 2022.156 2022.159 2022.162 2022.164 2022.164 2022.170
#> [257] 2022.173 2022.175 2022.178 2022.181 2022.184 2022.186 2022.189 2022.192
#> [265] 2022.195 2022.197 2022.200 2022.203 2022.205 2022.208 2022.211 2022.214
#> [273] 2022.216 2022.219 2022.222 2022.225 2022.227 2022.230 2022.233 2022.236
#> [281] 2022.238 2022.241 2022.244 2022.247 2022.249 2022.249 2022.255 2022.258
#> [289] 2022.260 2022.263 2022.266 2022.268 2022.271 2022.274 2022.277 2022.279
#> [297] 2022.282 2022.285 2022.288 2022.290 2022.293 2022.296 2022.299 2022.301
#> [305] 2022.304 2022.307 2022.310 2022.312 2022.315 2022.318 2022.321 2022.323
#> [313] 2022.326 2022.329 2022.332 2022.332 2022.337 2022.340 2022.342 2022.345
#> [321] 2022.348 2022.351 2022.353 2022.356 2022.359 2022.362 2022.364 2022.367
#> [329] 2022.370 2022.373 2022.375 2022.378 2022.381 2022.384 2022.386 2022.389
#> [337] 2022.392 2022.395 2022.397 2022.400 2022.403 2022.405 2022.408 2022.411
#> [345] 2022.414 2022.416 2022.416 2022.422 2022.425 2022.427 2022.430 2022.433
#> [353] 2022.436 2022.438 2022.441 2022.444 2022.447 2022.449 2022.452 2022.455
#> [361] 2022.458 2022.460 2022.463 2022.466 2022.468 2022.471
#>